Belt and suspender solution, this time actually fixing the bug.
A  trap rm with a relative path is dangerous and in this case
would not even have worked, since the problematic git command happens
after a "cd $repo", so a "rm -rf $repo" inside of $repo won't be doing
much.

Resolve $repo to an absolute path first and feed that to the trap.